refactor: 바운디드 컨텍스트 기준으로 패키지 수정 - 'eat_in_order', 'delivery_order', 'take_out_order' 에 각각 알맞는 OrderStatus 추가 - 'DIP' 를 사용하도록 구성 (application / domain / infra) - 테스트코드에서 import 문 알맞게 수정
