+++ src/main/java/kitchenpos/delivery_order/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.delivery_order.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
