+++ src/main/java/kitchenpos/delivery_order/infra/JpaDeliveryOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.delivery_order.domain;
+package kitchenpos.delivery_order.infra;
 
+import kitchenpos.delivery_order.domain.DeliveryOrder;
+import kitchenpos.delivery_order.domain.DeliveryOrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
