+++ src/main/java/kitchenpos/eat_in_order/domain/EatInOrder.java
@@ -2,7 +2,6 @@ package kitchenpos.eat_in_order.domain;
 
 import jakarta.persistence.*;
 import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderType;
 import kitchenpos.order_table.domain.OrderTable;
