+++ src/main/java/kitchenpos/eat_in_order/infra/JpaEatInOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.eat_in_order.domain;
+package kitchenpos.eat_in_order.infra;
 
+import kitchenpos.eat_in_order.domain.EatInOrder;
+import kitchenpos.eat_in_order.domain.EatInOrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
