+++ src/main/java/kitchenpos/take_out_order/application/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.take_out_order.infra;
+package kitchenpos.take_out_order.application;
 
 import java.math.BigDecimal;
 import java.util.UUID;
