+++ src/main/java/kitchenpos/take_out_order/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.take_out_order.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
