+++ src/main/java/kitchenpos/take_out_order/domain/TakeOutOrder.java
@@ -2,9 +2,7 @@ package kitchenpos.take_out_order.domain;
 
 import jakarta.persistence.*;
 import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderType;
-import kitchenpos.order_table.domain.OrderTable;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -30,10 +28,10 @@ public class TakeOutOrder {
 
     @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
     @JoinColumn(
-        name = "order_id",
-        nullable = false,
-        columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_order_line_item_to_orders")
+            name = "order_id",
+            nullable = false,
+            columnDefinition = "binary(16)",
+            foreignKey = @ForeignKey(name = "fk_order_line_item_to_orders")
     )
     private List<OrderLineItem> orderLineItems;
